Use file to initialize map

+++ src/main/java/ir/ac/kntu/gamelogic/models/tanks/ArmoredLuckyTank.java
@@ -4,7 +4,7 @@ import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.image.Image;
 
 public class ArmoredLuckyTank extends EnemyTank {
-    public ArmoredLuckyTank(int x, int y) {
+    public ArmoredLuckyTank(double x, double y) {
         super(x, y);
         health = 2;
         damage = 1;
